requirements.txt
# Core
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0

# Optional & Recommended
psycopg2-binary>=2.9       # PostgreSQL support
python-decouple>=3.8       # Manage secrets via .env
drf-yasg>=1.21             # Swagger/OpenAPI docs

# For cross-origin requests (e.g., if you use React or external frontend)
django-cors-headers>=4.3

# For JWT auth (optional)
djangorestframework-simplejwt>=5.3
